+++ sassdoc-webpack-plugin.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const fs = require('fs');
+const fs = require('fs/promises');
 const path = require('path');
 
 const yaml = require('js-yaml');