+++ scripts/compile-sprites.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const fs = require('fs');
+const fs = require('fs/promises');
 const path = require('path');
 
 const chalk = require('chalk');